requirements.txt
google-cloud-bigquery==1.8.1
google-cloud-storage==1.13.2
grpcio==1.18.0
requests-toolbelt==0.8.0
Werkzeug==0.14.1
Flask==1.0.2
GoogleAppEngineCloudStorageClient==1.9.15.0
Jinja2==2.10
google-api-core==1.7.0
requests==2.21.0